Pyrite -- an attempt to combine the best of Git an Mercurial
Currently the pyrite __init__ file has a starting list of the commands to be supported.  This is a fuctional goal for the first "release."  Once functionality is in place, try to use libgit to not have to shell out to do commands.  Then slowly strip away non-performance critical C code.  Leaving a user-friendly, cross platform, extensible and easy to set up SCM.