Templates/Styles.
Add templates and styles ability. Update log command to use it.

Templates and Styles are strings that contain properties enclosed in {}.  If
these {} have valid properties they can retrieve data from a dataset. The
dataset is a dictionary, probably containing values from the Repo's public
enumeration of commit attributes.  Certain properties of the Repo can also be
retrieved.

Asside from specifying what data you are looking for, you can also use a
formatter.  A formatter follows a property and is separated by a | symbol.

An example of a property/formatter may look like this...
{COMMIT_DATE|isodate}

where COMMIT_DATE is the property and isodate is the formatter.

"compile" must be called on the template class to prepare it to display data.
The compile method returns a bitfield of any properties found in the
template or style that is suitable for passing in to Repo.get_history or
Repo.show etc.

Signed-off-by: Govind Salinas <blix@sophiasuchtig.com>