checkin: Fix issues with amended checking with paths
amended checkin was not properly checking in the previous commits diff
plus only things on the command line.  This change applies the previous
diff to the index and then adds any files on the command line to the index.

In the event there are no paths added, everything is added as is normal
in pyrite.

Also cleaned up the pre-populated log message for amended checkins so that
it wouldn't have an extra line.

Signed-off-by: Govind Salinas <blix@sophiasuchtig.com>