Repo.add_files(): Allow to add a file from outside the repo dir.
A Repo object can technically have a cwd that is outside the git repo.
However adding a file will fail since git calculates the path relative to
the cwd. Even absolute paths were not working.

This now assumes that if you are calling from outside the repo root you are
passing in a path relative to the repo root.

Signed-off-by: Govind Salinas <blix@sophiasuchtig.com>