Repository.py: Use line based parsing instead of null based.
Previously null separators were used to parse the output from git log and
friends.  However, the null output was not as consistant as it could be
and there were bugs in the previous implementation. This method of parsing
uses the raw format and expects a specific format.  It is less flexible
but it works better and does not require any additional patches.

Signed-off-by: Govind Salinas blix@sophiasuchtig.com