Tweak git log parsing and log templates/generation to be more accurate.
Added Repo.REFS to the commit dict which holds the refs from --decorate.
Additionally tweak the way we are getting parts of the commit info through
the log to be more accurate.  In order to get the same output as before,
the templates and template class were tweaked, and are now more flexible.

Signed-off-by: Govind Salinas blix@sophiasuchtig.com