cherry.py: Make it more difficult to mess up pick.
It was easy to accidentally pick a commit at the head of a branch when
looking for commits from that branch if you forgot the -i.  Now -p must
be used or an error will be generated, this prevents the user from trying
to apply a commit they didn't want.

Signed-off-by: Govind Salinas blix@sophiasuchtig.com