checkin: Fix amended checkin with files.
The apply method of the repository needs to close stdin or the subprocess
may not exit as expected.

Diff uses a generator.  A generator does not execute until the first
time next() is called, which meant that the diff/apply mechanisim
in checkin was not capturing the right info.

Fix newlines on repository.changed_files.

Make instructions on aborting saving more clear.

Signed-off-by: Govind Salinas <blix@sophiasuchtig.com>