pyt alter
Command for rebasing and rewriting commit history.  Analogous to
git rebase.  Most modes of operation are working including interactive
mode, picking, squashing and editing.  Merging has yet to be
implemented.

Rebasing based on another branch appears to work as well as rewriting
the current branches history.

Generic sequencer is included in this commit that makes this possible.
In order to preserve state, the sequencer is serialized to disk and
restored on next run.

Signed-off-by: Govind Salinas blix@sophiasuchtig.com