repository.py: make get_history()'s options saner.
get_history's options for specifying revisions were confusing.  The
first item would be treated specially and sometimes it would actually
be used as the "last" commit.  Now the second commit is always the last
item and defaults to HEAD.

The lingering question is whether we should always show (first, last] as
git does or [first, last].  The second seems to make more sense most
of the time, but not when doing something like "git log master..next"
At that point, do you really want to see the last commit of master?

Signed-off-by: Govind Salinas blix@sophiasuchtig.com