spell_revisions.py: Proof of concept addon.
This addon is nice in that it defines revision numbers for git.  Using
this you will be able to use simple numbers followed by a colon to
refer to commits.  It also has a bunch of other ways to refer to
commits and reflog entries.

Additionally, it hooks the log command and puts the revision numbers
in the log output.

Signed-off-by: Govind Salinas <blix@sophiasuchtig.com>